+++ changelog.md
@@ -135,3 +135,4 @@ Added overloads to support specifying the name of the Initial Catalog using the 
 ## Changes
 * The assembly now targets .NET 4.0 instead of .NET 3.5.
 * A warning is logged if the configured version of the SQL LocalDB Instance API to load instead of the latest version cannot be found.
+* Fixed an exception being thrown by instances of ```TemporarySqlLocalDbInstance``` being thrown if an exception was thrown when the newly created temporary SQL LocalDB instance failed to start.